+++ nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.app/src/main/java/nts/uk/ctx/bs/employee/app/find/jobtitle/sequence/SequenceMasterFinder.java
@@ -31,7 +31,7 @@ public class SequenceMasterFinder {
 	 *
 	 * @return the short
 	 */
-	public short findMaxOrder() {
+	public int findMaxOrder() {
 		return this.repository.findMaxOrder();
 	}
