+++ nts.uk/uk.com/com.ctx/bs/employee/nts.uk.ctx.bs.employee.dom/src/main/java/nts/uk/ctx/bs/employee/dom/jobtitle/info/SequenceMaster.java
@@ -23,7 +23,7 @@ public class SequenceMaster extends AggregateRoot {
 
 	/** The order. */
 	//並び順
-	private short order;
+	private int order;
 
 	/** The sequence code. */
 	//序列コード
